Allow int animation index & speed, better UI states system

+++ Assets/Scripts/DataTypes/Unity/LevelObj/Animation/Unity_ObjAnimation.cs
@@ -10,6 +10,6 @@
         /// </summary>
         public Unity_ObjAnimationFrame[] Frames { get; set; }
 
-        public byte? AnimSpeed { get; set; }
+        public int? AnimSpeed { get; set; }
     }
 }
